Move IssueDetailsGetter to it's own file
