+++ src/bug-filing/common/issue-details-getter.ts
@@ -1,0 +1,8 @@
+import { EnvironmentInfo } from '../../common/environment-info-provider';
+import { CreateIssueDetailsTextData } from '../../common/types/create-issue-details-text-data';
+import { IssueUrlCreationUtils } from './issue-filing-url-string-utils';
+export type IssueDetailsGetter = (
+    stringUtils: IssueUrlCreationUtils,
+    environmentInfo: EnvironmentInfo,
+    data: CreateIssueDetailsTextData,
+) => string;
