+++ src/bug-filing/github/create-github-bug-filing-url.ts
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { getIssueDetailsMarkdown, IssueDetailsGetter } from '../common/get-issue-details-markdown';
+import { getIssueDetailsMarkdown } from '../common/get-issue-details-markdown';
+import { IssueDetailsGetter } from '../common/issue-details-getter';
 import { EnvironmentInfo } from './../../common/environment-info-provider';
 import { CreateIssueDetailsTextData } from './../../common/types/create-issue-details-text-data';
 import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from './../common/issue-filing-url-string-utils';
 import { GitHubBugFilingSettings } from './github-bug-filing-service';
 
-function buildTitle(stringUtils: IssueUrlCreationUtils, data: CreateIssueDetailsTextData): string {
+export function buildTitle(stringUtils: IssueUrlCreationUtils, data: CreateIssueDetailsTextData): string {
     const standardTags = stringUtils.standardizeTags(data);
     let prefix = standardTags.join(',');
     if (prefix.length > 0) {
